+++ stac_fastapi/elasticsearch/stac_fastapi/elasticsearch/core.py
@@ -368,12 +368,12 @@ class CoreClient(AsyncBaseCoreClient):
             base_args["sortby"] = sort_param
 
         if filter:
-            if filter_lang == "cql2-text":
-                base_args["filter-lang"] = "cql2-json"
-                base_args["filter"] = orjson.loads(to_cql2(parse_cql2_text(filter)))
-            else:
+            if filter_lang == "cql2-json":
                 base_args["filter-lang"] = "cql2-json"
                 base_args["filter"] = orjson.loads(unquote_plus(filter))
+            else:
+                base_args["filter-lang"] = "cql2-json"
+                base_args["filter"] = orjson.loads(to_cql2(parse_cql2_text(filter)))
 
         if fields:
             includes = set()